Small Changes
Small quality of life changes, all commands working. Made a dumb mistake in the data file, which caused the text to be formatted in a way where it took 40 lines or so. Took way to long to figure out what I'd done haha. Just working on another command or 2 and then just organising my code better