Small changes
Just changing the seller commands to make it buttons, might change it back, just deciding. Otherwise I'll just use the embed.py file and it'll be shorter either way. Just gotta work on the /ts command and get it into the database, work on the sellers command and make a couple commands embeds